Add SMML Documentation First Draft
Working from the notes I took when we met with Mr Jaworski, I
extrapolated the HTML-like concept of SMML to a fully-fledged
documentation, outlining the specifications for the various data types
and objects involved. This document is a work in progress, with some of
the specifications up for discussion and most of the phrasing,
structure, and explanations open to revision when passed under another
pair of eyes. Nonetheless, in its current state, I feel that this
document is ready to be called a "first draft."