Update to Frequency-Matching Experiment
01. Moved database files to filename "PitchData" to avoid confusion
02. Created a Python3 equivalent of the extant PitchFreq.js file

These changes are in preparation of the eventual packaging of the many
SMML-associated Python3 files into a single 'smmlPy' package. This
process will likely be the final stage of development, once all the
functions are defined and tested -- the final stage before releasing
SMML for public use (hopefully).

The creation of this package is how I (Evan) intend to respond to the
glut of Python3-NodeJS mixture going on in the repository. As python is
more commonly used for data processing, the Python3 package will define
an SMML class with various data processing functions. Any NodeJS in this
repository is primarily for for displaying SMML (which is not a question
python is meant to address).